chore: put all shapes to util/shapes
